+++ include/Modem/Modem.hpp
@@ -25,7 +25,7 @@ class Modem {
     Modem(int freqc, int freqs, double Timeb, int t = MODEM_BPSK);
     ~Modem();
     Eigen::RowVectorXd modulate(const Eigen::RowVectorXi &c);
-    Eigen::RowVectorXi demodulate(const Eigen::RowVectorXd &s);
+    Eigen::RowVectorXd demodulate(const Eigen::RowVectorXd &s);
     int getL();
 };
